rename utils.ts -> date.ts
