+++ src/calendar-range/calendar-range.tsx
@@ -1,6 +1,6 @@
 import { c, useEffect, useMemo, useState, type Host } from "atomico";
 import { PlainDate } from "../utils/temporal.js";
-import { inRange } from "../utils/utils.js";
+import { inRange } from "../utils/date.js";
 import { useDateRangeProp } from "../utils/hooks.js";
 import { CalendarBase, styles, props } from "../calendar-base/calendar-base.js";
 import { useCalendarBase } from "../calendar-base/useCalendarBase.js";
